+++ Questionable/Windows/JournalComponents/QuestJournalUtils.cs
@@ -196,7 +196,9 @@ internal sealed class QuestJournalUtils
 
         if (ImGui.Checkbox(_L("Show only Available Quests"), ref journalUi.Filter.AvailableOnly) ||
             ImGui.Checkbox(_L("Hide Quests Without Path"), ref journalUi.Filter.HideNoPaths) ||
-            ImGui.Checkbox(_L("Hide Completed Quests"), ref journalUi.Filter.HideCompleted))
+            ImGui.Checkbox(_L("Hide Completed Quests"), ref journalUi.Filter.HideCompleted) ||
+            ImGui.Checkbox(_L("Hide Unobtainable Quests"), ref journalUi.Filter.HideUnobtainable) ||
+            ImGui.Checkbox(_L("Hide Repeatable Quests"), ref journalUi.Filter.HideRepeatable))
         {
             journalUi.UpdateFilter();
         }
@@ -294,6 +296,8 @@ internal sealed class QuestJournalUtils
                 iconOverride = questInfo.InvalidIcon;
             if (icon is FontAwesomeIcon.Check)
                 iconOverride = questInfo.CompleteIcon;
+            if (questInfo.IsRepeatable && iconOverride % 10 < 2)
+                iconOverride = iconOverride + (2 - iconOverride % 10);
             if (icon is FontAwesomeIcon.QuestionCircle)
                 iconOverride = QuestionIcon;
         }